Retire use of O_EXEC for directories on FreeBSD
As of FreeBSD 11.4 and FreeBSD 12.2, O_SEARCH is now provided. While it is
defined as O_EXEC, it is better to retire the notion that O_EXEC will get
the job done in case that changes in the future. This is also consistent
with other platforms that provide O_PATH/O_SEARCH.