RAR5 reader: fixed out of bounds read in some files
The main command code decoding loop uses a simple range check on each
iteration to ensure that the unpacker will not read the memory beyond
its buffer.

If the extracted command code is a repetition command code, the unpacker
performs additional extraction of parameters that describe this
repetition code. This parameter extraction code performs stream reads on
its own, and there is a possibility that it will read more bytes than
it's available, before the range check from the main loop will have a
chance to verify the offset bounds.

This commit adds additional range check to a helper function that is
used inside various places during command code extraction algorithm, to
make sure that the stream bounds are respected also in the distance
parameter extraction code.

This commit contains a test with OSSFuzz sample #30448.