RAR5 reader: add more checks for invalid extraction parameters
Some specially crafted files declare invalid extraction parameters that
can confuse the RAR5 reader.

One of the arguments is the declared window size parameter that the
archive file can declare for each file stored in the archive. Some
crafted files declare window size equal to 0, which is clearly wrong.

RAR5 reader checks for invalid window size declaration. But if the
archive declares use of multiple volumes, then some sanity checks were
skipped due to wrong sequence of those checks.

This commit changes the sequence in which the checks are applied. Also
the check for invalid window size is changed, so it's checking the
actual value used later in the extractor, instead of a temporary
variable.

Another check for invalid window size argument is added into the actual
decompression function. This check is not crucial, but should act as a
second stage guard if another bug of this type would arise.

This commit also contains OSSFuzz sample #30459.